refactor: merge properties into combinatorics
